Se quito el selector de estrellas desde la card y se reemplazo por el metodo render estrellas que se usaba en el ProfesionalDetail
